refactor(checker): split engine_policy.go by concern (INV-0011 B1)
Three concerns were cohabiting a 1,324-line file alongside file-rule
evaluation. Extract them verbatim:

  engine_settings.go           211  setting-rule evaluation + remediation
  engine_branch_protection.go  224  rulesets-API evaluation
  engine_pr.go                 372  PR discovery, file sync, create/refresh
  engine_policy.go             562  file-rule evaluation (what remains)

Pure move, verified mechanically rather than by eye: the four files hold
the same 35 top-level declarations as the original, and comparing every
non-comment, non-import line as a multiset gives 942 lines before and
942 after with nothing appearing on only one side. No signature, no
statement, and no test changed.
