+++ internal/checker/engine_policy.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"log/slog"
 	"strings"
-	"time"
 
 	"gopkg.in/yaml.v3"
 
@@ -13,7 +12,6 @@ import (
 	"github.com/donaldgifford/repo-guardian/internal/metrics"
 	"github.com/donaldgifford/repo-guardian/internal/policy"
 	"github.com/donaldgifford/repo-guardian/internal/reconciler"
-	tmpl "github.com/donaldgifford/repo-guardian/internal/template"
 )
 
 // checkRepoWithPolicy runs policy-based evaluation for a repository.
@@ -511,26 +509,6 @@ func findExistingFile(
 	return "", nil
 }
 
-func hasExistingPRForPolicy(openPRs []*ghclient.PullRequest, rule *policy.FileRuleConfig) bool {
-	if rule.PR == nil {
-		return false
-	}
-
-	for _, pr := range openPRs {
-		titleLower := strings.ToLower(pr.Title)
-		branchLower := strings.ToLower(pr.Head)
-
-		for _, term := range rule.PR.SearchTerms {
-			termLower := strings.ToLower(term)
-			if strings.Contains(titleLower, termLower) || strings.Contains(branchLower, termLower) {
-				return true
-			}
-		}
-	}
-
-	return false
-}
-
 func isYAMLFile(path string) bool {
 	lower := strings.ToLower(path)
 
@@ -556,746 +534,6 @@ func yamlSemanticallyEqual(a, b string) (bool, error) {
 // yamlDeepEqual compares two parsed YAML values for equality.
 func yamlDeepEqual(a, b any) bool {
 	return fmt.Sprintf("%v", a) == fmt.Sprintf("%v", b)
-}
-
-// createOrUpdatePRFromPolicy creates or updates a PR for actionable policy rules.
-func (e *Engine) createOrUpdatePRFromPolicy(
-	ctx context.Context,
-	client ghclient.Client,
-	owner, repo, defaultBranch string,
-	actionable []policy.FileRuleConfig,
-	openPRs []*ghclient.PullRequest,
-	gate *gateEvaluator,
-) error {
-	log := e.logger.With("owner", owner, "repo", repo)
-
-	branchSHA, err := client.GetBranchSHA(ctx, owner, repo, BranchName)
-	if err != nil {
-		return fmt.Errorf("checking for existing branch: %w", err)
-	}
-
-	existingPR := findOurPR(openPRs)
-
-	if branchSHA != "" && existingPR == nil {
-		log.Info("deleting stale branch from previously closed PR")
-
-		if err := client.DeleteBranch(ctx, owner, repo, BranchName); err != nil {
-			return fmt.Errorf("deleting stale branch: %w", err)
-		}
-
-		branchSHA = ""
-	}
-
-	baseSHA, err := client.GetBranchSHA(ctx, owner, repo, defaultBranch)
-	if err != nil {
-		return fmt.Errorf("getting default branch SHA: %w", err)
-	}
-
-	if baseSHA == "" {
-		return fmt.Errorf("default branch %s has no SHA", defaultBranch)
-	}
-
-	if branchSHA == "" {
-		if err := client.CreateBranch(ctx, owner, repo, BranchName, baseSHA); err != nil {
-			return fmt.Errorf("creating branch: %w", err)
-		}
-
-		log.Info("created branch", "branch", BranchName)
-	}
-
-	// WARN: This loop syncs file content to the reconcile branch but does not
-	//   re-base the branch onto the current default-branch HEAD. If the PR sits
-	//   open while default advances, the branch's base SHA goes stale. Two
-	//   risks if auto-merge is enabled on the repo-guardian PR:
-	//   (1) base drift — usually safe (diff is "add this file") but loses
-	//       linear-history aesthetic; (2) content drift — if someone manually
-	//       writes a different version of the file to default in the gap, the
-	//       file rule's existence check stops flagging it, this loop no-ops,
-	//       and a squash-merge of the stale branch overwrites the manual
-	//       edit. Mitigations to consider: rebase the branch onto current
-	//       default before reconcile, close+reopen PRs older than N days,
-	//       or recommend operators don't enable auto-merge on
-	//       repo-guardian/* branches. See conversation in PR #71.
-	now := time.Now().UTC().Format(time.RFC3339)
-
-	if err := e.syncActionableFiles(ctx, log, client, owner, repo, defaultBranch, now, actionable); err != nil {
-		return err
-	}
-
-	if existingPR == nil {
-		return e.createNewPolicyPR(ctx, log, client, owner, repo, defaultBranch, now, actionable)
-	}
-
-	// IMPL-0013 Phase 3: existing PR is being updated. Remove orphan
-	// files (rules that were in a previous version of the PR but are
-	// now satisfied on the default branch) and refresh the PR body.
-	orphans := discoverOrphans(ctx, log, client, e.policy.FileRules, actionable, owner, repo)
-
-	var removedOrphans []string
-
-	if len(orphans) > 0 {
-		log.Info("cleaning up orphan files", "count", len(orphans))
-		removedOrphans = cleanupOrphans(ctx, log, client, orphans, owner, repo)
-	}
-
-	// IMPL-0019 Phase 2: inverse orphans — absent rules that stopped being
-	// actionable and whose forbidden file the branch deleted must be
-	// restored so the PR stops proposing the deletion.
-	restoredRules := restoreInverseOrphans(ctx, log, client, e.policy.FileRules, actionable, owner, repo)
-
-	if err := e.refreshPolicyPR(ctx, log, client, owner, repo, defaultBranch, existingPR, actionable, now); err != nil {
-		log.Warn("PR body refresh failed; PR text may be stale until next sweep", "err", err)
-	}
-
-	// IMPL-0013 Phase 4: sticky reconcile-log comment with per-rule
-	// status. Best-effort — failures don't abort the sweep.
-	events := buildReconcileLogEvents(e.policy.FileRules, actionable, removedOrphans, restoredRules, gate)
-	upsertReconcileLog(ctx, log, client, owner, repo, existingPR, events)
-
-	metrics.PRsUpdatedTotal.WithLabelValues(owner).Inc()
-	log.Info("updated existing PR", "pr_number", existingPR.Number)
-
-	return nil
-}
-
-// syncActionableFiles renders and commits the template content for
-// every actionable rule. Extracted from createOrUpdatePRFromPolicy
-// to keep cyclomatic complexity under the gocyclo threshold.
-func (e *Engine) syncActionableFiles(
-	ctx context.Context,
-	log *slog.Logger,
-	client ghclient.Client,
-	owner, repo, defaultBranch, now string,
-	actionable []policy.FileRuleConfig,
-) error {
-	for i := range actionable {
-		r := &actionable[i]
-
-		if r.CheckMode() == policy.CheckAbsent {
-			if err := e.removeForbiddenFiles(ctx, log, client, owner, repo, r); err != nil {
-				return err
-			}
-
-			continue
-		}
-
-		compiled, err := e.templates.Get(r.Template)
-		if err != nil {
-			return fmt.Errorf("getting template for %s: %w", r.Name, err)
-		}
-
-		content, err := compiled.Render(tmpl.FileVars{
-			Common: tmpl.Common{
-				Owner:         owner,
-				Repo:          repo,
-				DefaultBranch: defaultBranch,
-				Date:          now,
-			},
-			Rule: tmpl.Rule{Name: r.Name, Target: r.Target},
-			Org:  owner,
-		})
-		if err != nil {
-			return fmt.Errorf("rendering template for %s: %w", r.Name, err)
-		}
-
-		msg := fmt.Sprintf("chore: add %s", r.Target)
-
-		if err := client.CreateOrUpdateFile(ctx, owner, repo, BranchName, r.Target, content, msg); err != nil {
-			return fmt.Errorf("creating file %s: %w", r.Target, err)
-		}
-
-		log.Info("added file", "path", r.Target)
-	}
-
-	return nil
-}
-
-// removeForbiddenFiles deletes every path of an absent-mode rule that is
-// present on the reconcile branch, mirroring the INV-0003 three-branch
-// idempotency contract: probe the reconcile branch first, delete with the
-// blob SHA when present, skip when already absent so a second sweep is a
-// no-op. A path that is already gone is not an error — the branch has
-// converged for that path (IMPL-0019 Phase 2 task 2.1).
-func (*Engine) removeForbiddenFiles(
-	ctx context.Context,
-	log *slog.Logger,
-	client ghclient.Client,
-	owner, repo string,
-	r *policy.FileRuleConfig,
-) error {
-	for _, path := range r.Paths {
-		sha, exists, err := client.GetContentsOnBranch(ctx, owner, repo, path, BranchName)
-		if err != nil {
-			return fmt.Errorf("checking %s on branch for absent rule %q: %w", path, r.Name, err)
-		}
-
-		if !exists {
-			log.Debug("forbidden file already absent on branch", "path", path, "rule", r.Name)
-
-			continue
-		}
-
-		msg := fmt.Sprintf("chore: remove %s (forbidden by rule %q)", path, r.Name)
-
-		if err := client.DeleteFile(ctx, owner, repo, BranchName, path, sha, msg); err != nil {
-			return fmt.Errorf("deleting %s for absent rule %q: %w", path, r.Name, err)
-		}
-
-		log.Info("removed forbidden file", "path", path, "rule", r.Name)
-	}
-
-	return nil
-}
-
-// refreshPolicyPR re-renders the title and body for the current
-// actionable set and updates the PR if either drifted from the
-// in-flight values. No-op when both match — avoids API churn on
-// reconcile passes where nothing changed.
-func (e *Engine) refreshPolicyPR(
-	ctx context.Context,
-	log *slog.Logger,
-	client ghclient.Client,
-	owner, repo, defaultBranch string,
-	existingPR *ghclient.PullRequest,
-	actionable []policy.FileRuleConfig,
-	now string,
-) error {
-	fallbackBody := buildPRBodyFromPolicy(actionable)
-	vars := buildPRVars(owner, repo, defaultBranch, now, actionable)
-
-	var (
-		rendered renderedPR
-		err      error
-	)
-
-	switch len(actionable) {
-	case 1:
-		rendered, err = e.resolveAndRenderRulePR(log, &actionable[0], &vars, PRTitle, fallbackBody)
-	default:
-		rendered, err = e.resolveAndRenderBundlePR(log, actionable, &vars, PRTitle, fallbackBody)
-	}
-
-	if err != nil {
-		return fmt.Errorf("resolving PR template: %w", err)
-	}
-
-	if rendered.Title == existingPR.Title {
-		// Body is not exposed on the PR struct today; conservatively
-		// always issue the patch when the title is stable but the
-		// actionable set changed. The cost is one PATCH per sweep on
-		// touched repos, which fits inside the per-installation rate
-		// budget.
-		_ = rendered.Body
-	}
-
-	if err := client.UpdatePullRequest(ctx, owner, repo, existingPR.Number, rendered.Title, rendered.Body); err != nil {
-		return fmt.Errorf("update PR #%d: %w", existingPR.Number, err)
-	}
-
-	return nil
-}
-
-// createNewPolicyPR resolves PR templates, renders title/body/labels,
-// creates the PR, and applies labels. Extracted from
-// createOrUpdatePRFromPolicy to keep cyclomatic complexity in check.
-func (e *Engine) createNewPolicyPR(
-	ctx context.Context,
-	log *slog.Logger,
-	client ghclient.Client,
-	owner, repo, defaultBranch, now string,
-	actionable []policy.FileRuleConfig,
-) error {
-	fallbackBody := buildPRBodyFromPolicy(actionable)
-	vars := buildPRVars(owner, repo, defaultBranch, now, actionable)
-
-	var (
-		rendered renderedPR
-		err      error
-	)
-
-	switch len(actionable) {
-	case 1:
-		rendered, err = e.resolveAndRenderRulePR(log, &actionable[0], &vars, PRTitle, fallbackBody)
-	default:
-		rendered, err = e.resolveAndRenderBundlePR(log, actionable, &vars, PRTitle, fallbackBody)
-	}
-
-	if err != nil {
-		return fmt.Errorf("resolving PR template: %w", err)
-	}
-
-	pr, err := client.CreatePullRequest(ctx, owner, repo, rendered.Title, rendered.Body, BranchName, defaultBranch)
-	if err != nil {
-		return fmt.Errorf("creating PR: %w", err)
-	}
-
-	if err := client.AddLabelsToPR(ctx, owner, repo, pr.Number, rendered.Labels); err != nil {
-		log.Warn("adding labels to PR failed; continuing", "pr_number", pr.Number, "err", err)
-	}
-
-	metrics.PRsCreatedTotal.WithLabelValues(owner).Inc()
-	log.Info("created PR", "pr_number", pr.Number)
-
-	return nil
-}
-
-func buildPRBodyFromPolicy(actionable []policy.FileRuleConfig) string {
-	var added, removed []policy.FileRuleConfig
-
-	for i := range actionable {
-		if actionable[i].CheckMode() == policy.CheckAbsent {
-			removed = append(removed, actionable[i])
-		} else {
-			added = append(added, actionable[i])
-		}
-	}
-
-	var sb strings.Builder
-
-	sb.WriteString("## Repo Guardian — Configuration Drift\n\n")
-	sb.WriteString("This PR was automatically created by **repo-guardian** to bring this\n")
-	sb.WriteString("repository into line with the organization's file policy.\n\n")
-
-	if len(added) > 0 {
-		sb.WriteString("### Added Files\n\n")
-
-		for i := range added {
-			fmt.Fprintf(&sb, "- `%s` — %s\n", added[i].Target, added[i].Name)
-		}
-
-		// The CODEOWNERS placeholder note is only relevant when a file is
-		// being added; a removal-only PR must not mention it.
-		sb.WriteString("\n> **Note:** The CODEOWNERS file contains a placeholder (`@org/CHANGEME`).\n")
-		sb.WriteString("> Please replace it with your actual team before merging.\n\n")
-	}
-
-	if len(removed) > 0 {
-		sb.WriteString("### Removed Files\n\n")
-
-		for i := range removed {
-			for _, path := range removed[i].Paths {
-				fmt.Fprintf(&sb, "- `%s` — %s\n", path, removed[i].Name)
-			}
-		}
-
-		sb.WriteString("\n")
-	}
-
-	sb.WriteString("### What to do\n\n")
-	sb.WriteString("1. Review the changes and adjust for your team's needs.\n")
-	sb.WriteString("2. Merge when ready — these are sensible defaults, not one-size-fits-all.\n\n")
-	sb.WriteString("---\n")
-	sb.WriteString("*Automated by [repo-guardian](https://github.com/apps/repo-guardian). ")
-	sb.WriteString("Questions? Reach out in #platform-engineering.*\n")
-
-	return sb.String()
-}
-
-// evaluateSettingRules checks all setting rules against the repository.
-func (e *Engine) evaluateSettingRules(
-	ctx context.Context,
-	log *slog.Logger,
-	client ghclient.Client,
-	owner, repo string,
-) error {
-	if len(e.policy.SettingRules) == 0 {
-		return nil
-	}
-
-	ownerRepo := owner + "/" + repo
-	strict := strictMode(e.policy)
-
-	for i := range e.policy.SettingRules {
-		r := &e.policy.SettingRules[i]
-
-		if !r.IsEnabled() {
-			continue
-		}
-
-		ruleLog := log.With("setting_rule", r.Name, "property", r.Property)
-
-		if !ruleScopeAllows(r.Scope, owner, strict) {
-			ruleLog.Info("setting rule out of scope for org, skipping")
-			metrics.OutOfScopeTotal.WithLabelValues("rule", owner).Inc()
-
-			continue
-		}
-
-		if r.Ignore != nil && r.Ignore.Matches(ownerRepo) {
-			ruleLog.Info("repository matched per-rule ignore list, skipping setting rule")
-			metrics.IgnoredTotal.WithLabelValues("rule", owner).Inc()
-
-			continue
-		}
-
-		if err := e.evaluateSettingRule(ctx, ruleLog, client, owner, repo, r); err != nil {
-			return fmt.Errorf("evaluating setting rule %q: %w", r.Name, err)
-		}
-	}
-
-	return nil
-}
-
-// evaluateSettingRule checks a single setting rule against the repository.
-func (e *Engine) evaluateSettingRule(
-	ctx context.Context,
-	log *slog.Logger,
-	client ghclient.Client,
-	owner, repo string,
-	rule *policy.SettingRuleConfig,
-) error {
-	metrics.SettingsCheckedTotal.WithLabelValues(rule.Name, owner).Inc()
-
-	currentValue, err := e.getSettingValue(ctx, client, owner, repo, rule.Property)
-	if err != nil {
-		return fmt.Errorf("getting current value for %s: %w", rule.Property, err)
-	}
-
-	if settingMatches(currentValue, rule.Expected) {
-		log.Debug("setting matches expected value", "current", currentValue)
-		return nil
-	}
-
-	metrics.SettingsMismatchedTotal.WithLabelValues(rule.Name, owner).Inc()
-	log.Info("setting mismatch", "current", currentValue, "expected", rule.Expected)
-
-	if !rule.Remediate {
-		return nil
-	}
-
-	if e.dryRun {
-		log.Info("dry run: would remediate setting", "current", currentValue, "expected", rule.Expected)
-		return nil
-	}
-
-	if err := e.remediateSetting(ctx, log, client, owner, repo, rule); err != nil {
-		return fmt.Errorf("remediating %s: %w", rule.Property, err)
-	}
-
-	metrics.SettingsRemediatedTotal.WithLabelValues(rule.Name, owner).Inc()
-	log.Info("remediated setting", "property", rule.Property)
-
-	return nil
-}
-
-// getSettingValue reads the current value of a repository setting.
-func (*Engine) getSettingValue(
-	ctx context.Context,
-	client ghclient.Client,
-	owner, repo, property string,
-) (any, error) {
-	if property == policy.SettingVulnerabilityAlertsEnabled {
-		return client.GetVulnerabilityAlertsEnabled(ctx, owner, repo)
-	}
-
-	settings, err := client.GetRepoSettings(ctx, owner, repo)
-	if err != nil {
-		return nil, err
-	}
-
-	switch property {
-	case policy.SettingDefaultBranch:
-		return settings.DefaultBranch, nil
-	case policy.SettingHasIssues:
-		return settings.HasIssues, nil
-	case policy.SettingHasWiki:
-		return settings.HasWiki, nil
-	case policy.SettingDeleteBranchOnMerge:
-		return settings.DeleteBranchOnMerge, nil
-	case policy.SettingAllowMergeCommit:
-		return settings.AllowMergeCommit, nil
-	case policy.SettingAllowSquashMerge:
-		return settings.AllowSquashMerge, nil
-	case policy.SettingAllowRebaseMerge:
-		return settings.AllowRebaseMerge, nil
-	default:
-		return nil, fmt.Errorf("unsupported property: %s", property)
-	}
-}
-
-// settingMatches compares the current value against the expected value.
-func settingMatches(current, expected any) bool {
-	return fmt.Sprintf("%v", current) == fmt.Sprintf("%v", expected)
-}
-
-// remediateSetting applies the expected setting value via the GitHub API.
-func (*Engine) remediateSetting(
-	ctx context.Context,
-	_ *slog.Logger,
-	client ghclient.Client,
-	owner, repo string,
-	rule *policy.SettingRuleConfig,
-) error {
-	switch rule.Property {
-	case policy.SettingVulnerabilityAlertsEnabled:
-		expected, ok := rule.Expected.(bool)
-		if !ok {
-			return fmt.Errorf("expected bool for %s, got %T", policy.SettingVulnerabilityAlertsEnabled, rule.Expected)
-		}
-
-		if expected {
-			return client.EnableVulnerabilityAlerts(ctx, owner, repo)
-		}
-
-		return client.DisableVulnerabilityAlerts(ctx, owner, repo)
-
-	case policy.SettingDefaultBranch:
-		expected, ok := rule.Expected.(string)
-		if !ok {
-			return fmt.Errorf("expected string for %s, got %T", policy.SettingDefaultBranch, rule.Expected)
-		}
-
-		return client.UpdateRepository(ctx, owner, repo, &ghclient.RepoUpdateOpts{
-			DefaultBranch: &expected,
-		})
-
-	default:
-		return remediateBoolSetting(ctx, client, owner, repo, rule)
-	}
-}
-
-// remediateBoolSetting handles remediation for boolean repository settings.
-func remediateBoolSetting(
-	ctx context.Context,
-	client ghclient.Client,
-	owner, repo string,
-	rule *policy.SettingRuleConfig,
-) error {
-	expected, ok := rule.Expected.(bool)
-	if !ok {
-		return fmt.Errorf("expected bool for %s, got %T", rule.Property, rule.Expected)
-	}
-
-	opts := &ghclient.RepoUpdateOpts{}
-
-	switch rule.Property {
-	case policy.SettingHasIssues:
-		opts.HasIssues = &expected
-	case policy.SettingHasWiki:
-		opts.HasWiki = &expected
-	case policy.SettingDeleteBranchOnMerge:
-		opts.DeleteBranchOnMerge = &expected
-	case policy.SettingAllowMergeCommit:
-		opts.AllowMergeCommit = &expected
-	case policy.SettingAllowSquashMerge:
-		opts.AllowSquashMerge = &expected
-	case policy.SettingAllowRebaseMerge:
-		opts.AllowRebaseMerge = &expected
-	default:
-		return fmt.Errorf("unsupported bool property: %s", rule.Property)
-	}
-
-	return client.UpdateRepository(ctx, owner, repo, opts)
-}
-
-// evaluateBranchProtectionRules checks all branch protection rules against the repository.
-func (e *Engine) evaluateBranchProtectionRules(
-	ctx context.Context,
-	log *slog.Logger,
-	client ghclient.Client,
-	owner, repo string,
-) error {
-	if len(e.policy.BranchProtectionRules) == 0 {
-		return nil
-	}
-
-	ownerRepo := owner + "/" + repo
-	strict := strictMode(e.policy)
-
-	for i := range e.policy.BranchProtectionRules {
-		r := &e.policy.BranchProtectionRules[i]
-
-		if !r.IsEnabled() {
-			continue
-		}
-
-		ruleLog := log.With("bp_rule", r.Name, "branch", r.Branch)
-
-		if !ruleScopeAllows(r.Scope, owner, strict) {
-			ruleLog.Info("branch protection rule out of scope for org, skipping")
-			metrics.OutOfScopeTotal.WithLabelValues("rule", owner).Inc()
-
-			continue
-		}
-
-		if r.Ignore != nil && r.Ignore.Matches(ownerRepo) {
-			ruleLog.Info("repository matched per-rule ignore list, skipping branch protection rule")
-			metrics.IgnoredTotal.WithLabelValues("rule", owner).Inc()
-
-			continue
-		}
-
-		if err := e.evaluateBranchProtectionRule(ctx, ruleLog, client, owner, repo, r); err != nil {
-			return fmt.Errorf("evaluating branch protection rule %q: %w", r.Name, err)
-		}
-	}
-
-	return nil
-}
-
-// evaluateBranchProtectionRule checks a single branch protection rule.
-func (e *Engine) evaluateBranchProtectionRule(
-	ctx context.Context,
-	log *slog.Logger,
-	client ghclient.Client,
-	owner, repo string,
-	rule *policy.BranchProtectionRuleConfig,
-) error {
-	metrics.BranchProtectionCheckedTotal.WithLabelValues(rule.Name, owner).Inc()
-
-	// Check if the branch exists.
-	sha, err := client.GetBranchSHA(ctx, owner, repo, rule.Branch)
-	if err != nil {
-		return fmt.Errorf("checking branch %s: %w", rule.Branch, err)
-	}
-
-	if sha == "" {
-		log.Warn("branch does not exist, skipping branch protection rule")
-		return nil
-	}
-
-	// Fetch existing rulesets.
-	rulesets, err := client.ListRepositoryRulesets(ctx, owner, repo)
-	if err != nil {
-		return fmt.Errorf("listing rulesets: %w", err)
-	}
-
-	// Find a matching ruleset for this branch.
-	existing := findMatchingRuleset(rulesets, rule)
-
-	mismatches := compareBranchProtection(existing, rule)
-	if len(mismatches) == 0 {
-		log.Debug("branch protection matches expected configuration")
-		return nil
-	}
-
-	log.Info("branch protection mismatch", "mismatches", mismatches)
-
-	if !rule.Remediate {
-		return nil
-	}
-
-	if e.dryRun {
-		log.Info("dry run: would remediate branch protection", "mismatches", mismatches)
-		return nil
-	}
-
-	desired := buildDesiredRuleset(rule)
-
-	if existing != nil {
-		if _, err := client.UpdateRepositoryRuleset(ctx, owner, repo, existing.ID, desired); err != nil {
-			return fmt.Errorf("updating ruleset: %w", err)
-		}
-
-		log.Info("updated branch protection ruleset")
-	} else {
-		if _, err := client.CreateRepositoryRuleset(ctx, owner, repo, desired); err != nil {
-			return fmt.Errorf("creating ruleset: %w", err)
-		}
-
-		log.Info("created branch protection ruleset")
-	}
-
-	metrics.BranchProtectionRemediatedTotal.WithLabelValues(rule.Name, owner).Inc()
-
-	return nil
-}
-
-// findMatchingRuleset finds a ruleset that targets the same branch pattern.
-func findMatchingRuleset(
-	rulesets []*ghclient.Ruleset,
-	rule *policy.BranchProtectionRuleConfig,
-) *ghclient.Ruleset {
-	for _, rs := range rulesets {
-		if rs.Conditions == nil {
-			continue
-		}
-
-		for _, pattern := range rs.Conditions.IncludePatterns {
-			if pattern == rule.Branch || pattern == "refs/heads/"+rule.Branch {
-				return rs
-			}
-		}
-	}
-
-	return nil
-}
-
-// compareBranchProtection compares the existing ruleset against the desired config.
-// Returns a list of mismatch descriptions.
-func compareBranchProtection(
-	existing *ghclient.Ruleset,
-	rule *policy.BranchProtectionRuleConfig,
-) []string {
-	if existing == nil {
-		if rule.RequirePR || rule.RequireLinearHistory || len(rule.RequireStatusChecks) > 0 {
-			return []string{"no matching ruleset found"}
-		}
-
-		return nil
-	}
-
-	var mismatches []string
-
-	if rule.RequirePR && existing.RequirePullRequest == nil {
-		mismatches = append(mismatches, "pull request required but not configured")
-	}
-
-	if rule.RequirePR && existing.RequirePullRequest != nil {
-		pr := existing.RequirePullRequest
-
-		if pr.RequiredApprovals != rule.RequiredApprovals {
-			mismatches = append(mismatches, fmt.Sprintf(
-				"required_approvals: got %d, want %d",
-				pr.RequiredApprovals, rule.RequiredApprovals,
-			))
-		}
-
-		if pr.DismissStaleReviews != rule.DismissStaleReviews {
-			mismatches = append(mismatches, fmt.Sprintf(
-				"dismiss_stale_reviews: got %v, want %v",
-				pr.DismissStaleReviews, rule.DismissStaleReviews,
-			))
-		}
-	}
-
-	if rule.RequireLinearHistory != existing.RequireLinearHistory {
-		mismatches = append(mismatches, fmt.Sprintf(
-			"require_linear_history: got %v, want %v",
-			existing.RequireLinearHistory, rule.RequireLinearHistory,
-		))
-	}
-
-	return mismatches
-}
-
-// buildDesiredRuleset creates a Ruleset from the branch protection config.
-func buildDesiredRuleset(rule *policy.BranchProtectionRuleConfig) *ghclient.Ruleset {
-	rs := &ghclient.Ruleset{
-		Name:        "repo-guardian-" + rule.Name,
-		Enforcement: "active",
-		Target:      "branch",
-		Conditions: &ghclient.RulesetConditions{
-			IncludePatterns: []string{"refs/heads/" + rule.Branch},
-		},
-		RequireLinearHistory: rule.RequireLinearHistory,
-	}
-
-	if rule.RequirePR {
-		rs.RequirePullRequest = &ghclient.RulesetPullRequest{
-			RequiredApprovals:   rule.RequiredApprovals,
-			DismissStaleReviews: rule.DismissStaleReviews,
-		}
-	}
-
-	if len(rule.RequireStatusChecks) > 0 {
-		rs.RequireStatusChecks = &ghclient.RulesetStatusChecks{
-			RequiredChecks:     rule.RequireStatusChecks,
-			StrictStatusChecks: true,
-		}
-	}
-
-	return rs
 }
 
 // policyRuleNames extracts rule names from policy file rules.
